Eliminate superfluous consts.
The presence of these consts trigger warnings under -Wignored-qualifiers
and -Wextra modes.

C++ provides no facilities to ensure that the pointer returned from a
function is not modified.  For objects, this is safe, because messing
with a returned pointer will not alter the object's pointer:

	struct S
	{
		int const* const getA() {
			return &a;
		}

		int a;
	}

	S s;
	s.a = 42;

	int const* b = s.getA();	// yeah, this is fine
	b = 0;

	// s.a has not been modified

The reasons for this behavior, as far as I can tell, are:

1. b and s.a are two locations in memory that, prior to the b = 0
   assignment, have the same value: the same memory address.
2. In current versions of C++, the returned pointer is a value owned by
   the caller; the callee can't prevent the caller from manipulating that
   pointer to point somewhere else.

(There is probably some language in the C++ standard regarding permitted
const levels on rvalues, but I don't know where that is.)